Fix the error which would happen after running the manifest updater once
The error wasn't with String.trim() as I had originally thought, instead it was with the way I was overwriting the version file so the newer version could be written to it.
I was originally just setting the file's length to 0, assuming it would remove the contents of the file completely, leaving it ready to have the new version written to it. That isn't how rust's set_len() method works; I was already at offset 12 in the file, so running truncate there just left me with 12 null bytes.
The next time the manifest updater was run, it would open that file, grab the bytes inside, convert them to a string, then pass that off to the Version struct. Which would split the file up by periods and try to parse the substrings to numbers. The year substring would have 12 null bytes, then the year, thus causing the bug.